Update infect.pl
If a soldier targets a nurse or a doctor, and the action would result in death, the soldier rolls once more; this will most likely circumvent the death. This has drastically reduced the amount of friendly-fire-related deaths. In turn, this has increased the doctor/nurse population by quite a bit as well.